refactor(sport-cycling): relocate src/cycling/ + SOUL.md + skills/ → packages/sport-cycling/; create CONTEXT.md; three-category tool split per ADR-0004
Wave 3, commit 6 of issue #47.

Source moves (12 files):
- src/cycling/{feasibility,index,intervals-serializer,periodization,plan-builder,schemas,sport,templates,tools,zones}.ts → packages/sport-cycling/src/
- src/cycling/migrate-legacy-sections.ts → packages/sport-cycling/src/migrate.ts (rename for shorter import path: `@enduragent/sport-cycling/migrate`)
- src/cycling/binary.ts → packages/sport-cycling/src/binary.ts (intermediate; commit 7 moves it again to packages/cycling-coach/)

Asset moves:
- SOUL.md (root) → packages/sport-cycling/SOUL.md
- skills/{intervals-icu,periodization,race-prep,recovery,workout-design,zone-reference}.md → packages/sport-cycling/skills/

Test moves (8 files):
- 7 cycling tests → packages/sport-cycling/tests/
- tests/chronic-keyword-scan.test.ts → packages/core/tests/ (Core-owned scan logic per spec)
- tests/migration-integration.test.ts stays at root (moves to packages/cycling-coach/ in commit 7)

Three-category tool split (carryover from commit 3, locked in /grill-with-docs Q1):
Per ADR-0004, intervals tools split by variation:
- New `packages/core/src/agent/intervals-tools.ts` exports `createPureCoreIntervalsTools(intervals)` (intervals_fetch_athlete, intervals_fetch_wellness, intervals_delete_workout) and `createCoreToolsWithSportConfig(intervals, activityTypes)` (intervals_fetch_activities, intervals_list_events).
- `packages/sport-cycling/src/tools.ts` slimmed to Pure-Sport: calculate_zones, build_plan_skeleton, assess_feasibility, get_sample_week, intervals_create_workout (still hardcoded `type: "Ride"`; future split to `intervals_create_ride_workout` is a sport-cycling internal change).
- `cyclingSport.tools(deps)` now composes four buckets: createMemoryTools + createPureCoreIntervalsTools + createCoreToolsWithSportConfig(deps.intervals, cyclingSport.intervalsActivityTypes) + createCyclingTools.

Package wiring:
- packages/sport-cycling/package.json gains exports map: ".", "./binary", "./migrate".
- packages/sport-cycling/src/index.ts re-exports cyclingSport, CYCLING_VOCABULARY, cyclingBinary, migrateCyclingLegacySections.
- New packages/sport-cycling/tsconfig.json (mirrors Core's).
- New packages/sport-cycling/CONTEXT.md per spec template.
- Root package.json: + `@enduragent/sport-cycling` workspace dep, + `build:sport-cycling` and `build:packages` scripts; `build` and `check` use `build:packages`.
- Root tsconfig.json paths: + @enduragent/sport-cycling, /binary, /migrate subpath maps to dist .d.ts.
- src/index.ts: cyclingSport, cyclingBinary, migrateCyclingLegacySections imports redirect from src/cycling/* to @enduragent/sport-cycling.

Cycling cross-package test imports (transient until commit 7's binary move):
- packages/core/tests/{compaction,retry-loop-codex,setup-merge,setup.secret-ref,codex-setup-flow}.test.ts: `../../../src/cycling/*` → `@enduragent/sport-cycling` (or /binary subpath). Now that sport-cycling is a real package, the imports are clean even though the tests are in a sibling package.
- tests/migration-integration.test.ts: same redirect.

PROJECT_ROOT bug fix: `packages/sport-cycling/src/sport.ts` had `resolve(import.meta.dirname, "../..")` which after the move resolved to `packages/` (wrong). Renamed to `PACKAGE_ROOT = resolve(import.meta.dirname, "..")` — `packages/sport-cycling/`. SOUL.md and skills/ load correctly both from src/ (dev) and dist/ (built/published).

Verification:
- `find src/cycling -type f` → empty (directory removed).
- `find packages/sport-cycling/src -type f -name "*.ts" | wc -l` → 12 (≥ 10 required).
- `packages/sport-cycling/SOUL.md` exists; root `SOUL.md` does not.
- `find packages/sport-cycling/skills -name "*.md" | wc -l` → 6.
- `packages/sport-cycling/CONTEXT.md` exists with canonical glossary.
- `grep -r "../../src/" packages/sport-cycling/` → empty.
- packages/core build clean. packages/sport-cycling build clean. root build clean.
- root lint clean.
- pnpm test: 317 passed, 1 skipped, 0 failed.
