+++ packages/sport-cycling/src/intervals-serializer.ts
