+++ packages/sport-cycling/src/templates.ts
