+++ src/index.ts
@@ -1,8 +1,7 @@
 #!/usr/bin/env node
 import { runBinary } from "@enduragent/core";
-import { cyclingSport } from "./cycling/sport.js";
-import { cyclingBinary } from "./cycling/binary.js";
-import { migrateCyclingLegacySections } from "./cycling/migrate-legacy-sections.js";
+import { cyclingSport, cyclingBinary } from "@enduragent/sport-cycling";
+import { migrateCyclingLegacySections } from "@enduragent/sport-cycling/migrate";
 
 await runBinary(cyclingSport, cyclingBinary, {
   onStartup: (memory) => migrateCyclingLegacySections(memory),
